OOPed new mod items for automated localization
