+++ Content/Items/Accessories/RiskBadge.cs
@@ -2,6 +2,7 @@
 using Terraria;
 using Terraria.ID;
 using Terraria.ModLoader;
+using TerraTemp.Content.Base.Items;
 
 namespace TerraTemp.Content.Items.Accessories {
