+++ Content/Items/Base/TempItem.cs
@@ -1,16 +1,15 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Terraria;
 using Terraria.ModLoader;
 using TerraTemp.Custom;
 
-namespace TerraTemp.Content.Items {
+namespace TerraTemp.Content.Base.Items {
 
     /// <summary>
-    /// Abstract class that extends ModItem that has built in functionality adding Equipment that
-    /// changes temperature statistics.
+    /// Abstract class for any item that can affect a given player's temperature stats in ANY given
+    /// way. Tooltip is automatically handled, no matter the change.
     /// </summary>
-    public abstract class TempEquipment : ModItem {
+    public abstract class TempItem : ModItem {
 
         /// <summary>
         /// By how much this given item will change the player's Base Desired (Environmental) Temperature.
@@ -45,24 +44,29 @@ namespace TerraTemp.Content.Items {
         public override void ModifyTooltips(List<TooltipLine> tooltips) {
             string returnedLine = TempUtilities.CreateNewLineBasedOnStats(HeatComfortabilityChange, ColdComfortabilityChange, TemperatureResistanceChange, CriticalTemperatureChange, DesiredTemperatureChange, ClimateExtremityChange);
             if (returnedLine != null) {
+                TooltipLine newLine = new TooltipLine(mod, "TempAdditionalLine", returnedLine);
+
+                TooltipLine defenseLine = tooltips.FirstOrDefault(t => t.mod == "Terraria" && t.Name == "Defense");
+                TooltipLine buffDurationLine = tooltips.FirstOrDefault(t => t.mod == "Terraria" && t.Name == "BuffTime");
                 TooltipLine modifierLine = tooltips.FirstOrDefault(t => t.mod == "Terraria" && t.Name.Contains("Prefix"));
-                if (modifierLine != null) {
-                    tooltips.Insert(tooltips.IndexOf(modifierLine), new TooltipLine(mod, "TemperatureStatChange", returnedLine));
+                TooltipLine sellLine = tooltips.FirstOrDefault(tooltip => tooltip.mod == "Terraria" && (tooltip.Name == "Price" || tooltip.Name == "SpecialPrice"));
+
+                if (defenseLine != null) {
+                    tooltips.Insert(tooltips.IndexOf(defenseLine) + 1, newLine);
+                }
+                else if (buffDurationLine != null) {
+                    tooltips.Insert(tooltips.IndexOf(buffDurationLine), newLine);
+                }
+                else if (modifierLine != null) {
+                    tooltips.Insert(tooltips.IndexOf(modifierLine), newLine);
+                }
+                else if (sellLine != null) {
+                    tooltips.Insert(tooltips.IndexOf(sellLine), newLine);
                 }
                 else {
-                    tooltips.Add(new TooltipLine(mod, "TemperatureStatChange", returnedLine));
+                    tooltips.Add(newLine);
                 }
             }
-        }
-
-        public override void UpdateEquip(Player player) {
-            TempPlayer temperaturePlayer = player.GetTempPlayer();
-            temperaturePlayer.baseDesiredTemperature += DesiredTemperatureChange;
-            temperaturePlayer.comfortableHigh += HeatComfortabilityChange;
-            temperaturePlayer.comfortableLow += ColdComfortabilityChange;
-            temperaturePlayer.temperatureChangeResist += TemperatureResistanceChange;
-            temperaturePlayer.criticalRangeMaximum += CriticalTemperatureChange;
-            temperaturePlayer.climateExtremityValue += ClimateExtremityChange;
         }
     }
 }
