+++ Localization/en-US.lang
@@ -69,6 +69,9 @@ ItemName.BarbFish=Cherry Barb
 ItemName.Thermometer=Thermometer
 ItemTooltip.Thermometer=Displays several temperature related statistics
 
+#TerraTemp Potion Items/Tooltips
+ItemName.CoolingPotion=Cooling Potion
+
 #TerraTemp Buff Names/Tooltips
 BuffName.CoolingPotionBuff=Cooled
